moved import and export data transformations into datatype methods and created a base concept type from which concept types inherit, re #1541
